1.更新buildTool 2.新增native daemon 3.其他更新

+++ lib/MyUtil/src/main/java/cm/android/util/UserDataCleaner.java
@@ -95,7 +95,17 @@ public class UserDataCleaner {
      * 删除方法
      */
     private static void deleteFilesByDirectory(File directory) {
-        cm.java.util.IoUtil.deleteFile(directory);
+        if (directory == null) {
+            return;
+        }
+
+        File to = new File(directory.getAbsolutePath() + System.currentTimeMillis());
+        boolean flag = directory.renameTo(to);
+        if (flag) {
+            cm.java.util.IoUtil.deleteFile(to);
+        } else {
+            cm.java.util.IoUtil.deleteFile(directory);
+        }
 //        IoUtil.deleteFiles(directory);
     }
